Modulo equipos (tablas ram,almacenamiento,sistema Operativo, procesadores (Agregar y eliminar))

+++ Roles/admi/acciones.php
@@ -439,7 +439,99 @@ if($_SERVER['REQUEST_METHOD'] === 'GET'){ // aca hago la comprobacion si la peti
         insertTable($mysqli, $tabla , 'nom_formacion', $_POST['nameTipo']);
     }elseif($tabla === 'tip_periferico'){
         insertTable($mysqli, $tabla, 'nom_tip_periferico', $_POST['nameTipo']);
-    }elseif ($tabla === 'detalle_formacion'){
+    }elseif ($tabla === 'ram'){
+
+        $sql = "INSERT INTO ram (ramGB, tamaño_ram) VALUES (NULL, ?)";
+        $query = mysqli_prepare($mysqli, $sql);
+        $ok = mysqli_stmt_bind_param($query, 's', $_POST['memoriaRam']);
+        $ok = mysqli_stmt_execute($query);
+        mysqli_stmt_close($query);
+        $res;
+        if($ok){
+            $res = array (
+                'err' => false,
+                'status' => http_response_code(200),
+                'statusText' => 'Registro insertado con exito',
+            );
+        }else{     
+            $res = array (
+                'err' => true,
+                'status' => http_response_code(500),
+                'statusText' => 'No se puede insertar el registro',
+            );
+        }
+        echo json_encode($res);
+
+    }elseif ($tabla === 'tipo_sistema'){
+
+        $sql = "INSERT INTO tipo_sistema (id_tipo_sistema, nom_tipo_sistema) VALUES (NULL, ?)";
+        $query = mysqli_prepare($mysqli, $sql);
+        $ok = mysqli_stmt_bind_param($query, 's', $_POST['sistema_op']);
+        $ok = mysqli_stmt_execute($query);
+        mysqli_stmt_close($query);
+        $res;
+        if($ok){
+            $res = array (
+                'err' => false,
+                'status' => http_response_code(200),
+                'statusText' => 'Registro insertado con exito',
+            );
+        }else{     
+            $res = array (
+                'err' => true,
+                'status' => http_response_code(500),
+                'statusText' => 'No se puede insertar el registro',
+            );
+        }
+        echo json_encode($res);
+
+    }
+    
+    elseif ($tabla === 'almacenamiento'){
+        $sql = "INSERT INTO almacenamiento (id_almacena, tamaño_almacena) VALUES (NULL, ?)";
+        $query = mysqli_prepare($mysqli, $sql);
+        $ok = mysqli_stmt_bind_param($query, 's', $_POST['almacenamiento']);
+        $ok = mysqli_stmt_execute($query);
+        mysqli_stmt_close($query);
+        $res;
+        if($ok){
+            $res = array (
+                'err' => false,
+                'status' => http_response_code(200),
+                'statusText' => 'Registro insertado con exito',
+            );
+        }else{     
+            $res = array (
+                'err' => true,
+                'status' => http_response_code(500),
+                'statusText' => 'No se puede insertar el registro',
+            );
+        }
+        echo json_encode($res);
+
+    }elseif ($tabla === 'procesadores'){
+        $sql = "INSERT INTO procesadores (id_procesador, nom_procesador) VALUES (NULL, ?)";
+        $query = mysqli_prepare($mysqli, $sql);
+        $ok = mysqli_stmt_bind_param($query, 's', $_POST['nom_procesador']);
+        $ok = mysqli_stmt_execute($query);
+        mysqli_stmt_close($query);
+        $res;
+        if($ok){
+            $res = array (
+                'err' => false,
+                'status' => http_response_code(200),
+                'statusText' => 'Registro insertado con exito',
+            );
+        }else{     
+            $res = array (
+                'err' => true,
+                'status' => http_response_code(500),
+                'statusText' => 'No se puede insertar el registro',
+            );
+        }
+        echo json_encode($res);
+    }
+    elseif ($tabla === 'detalle_formacion'){
         $sql = "INSERT INTO detalle_formacion (id_detalle_formacion, id_formacion, num_ficha, id_ambiente) VALUES (NULL, ?, ?, ? )";
         $query = mysqli_prepare($mysqli, $sql);
         $ok = mysqli_stmt_bind_param($query, 'iii', $_POST['formacion'],$_POST['num_ficha'], $_POST['ambiente'] );
@@ -856,7 +948,7 @@ if($_SERVER['REQUEST_METHOD'] === 'GET'){ // aca hago la comprobacion si la peti
     $_DELETE = json_decode(file_get_contents('php://input'), true);
     $tabla = $_DELETE['tabla'];
     $id = $_DELETE['id'];
-    if($tabla !== 'dispositivo_electronico' && $tabla !== 'usuarios' && $tabla !== 'fichas' && $tabla !== 'compu_peris'){
+    if($tabla !== 'dispositivo_electronico' && $tabla !== 'usuarios' && $tabla !== 'fichas' && $tabla !== 'compu_peris' && $tabla !== 'ram' && $tabla !== 'almacenamiento' && $tabla !== 'tipo_sistema' && $tabla !== 'procesadores'){
         $sql = "DELETE from $tabla where id_$tabla = ?";
         $query = mysqli_prepare($mysqli , $sql);
         $ok = mysqli_stmt_bind_param($query, 's' , $id);
@@ -876,6 +968,58 @@ if($_SERVER['REQUEST_METHOD'] === 'GET'){ // aca hago la comprobacion si la peti
                 'statusText' => 'no se logro borrar el registro',
             );
         }
+        echo json_encode($res);
+    }
+    if($tabla === 'ram' ){
+        $sql = "DELETE from $tabla where ramGB = ? ";
+        $query = mysqli_prepare($mysqli , $sql);
+        $ok = mysqli_stmt_bind_param($query, 'i' , $id);
+        $ok = mysqli_stmt_execute($query);
+        mysqli_stmt_close($query);
+        $res = array (
+            'err' => false,
+            'status' => http_response_code(200),
+            'statusText' => 'Registro borrado con exito!',
+        );
+        echo json_encode($res);
+    }
+    if($tabla === 'almacenamiento' ){
+        $sql = "DELETE from $tabla where id_almacena = ? ";
+        $query = mysqli_prepare($mysqli , $sql);
+        $ok = mysqli_stmt_bind_param($query, 'i' , $id);
+        $ok = mysqli_stmt_execute($query);
+        mysqli_stmt_close($query);
+        $res = array (
+            'err' => false,
+            'status' => http_response_code(200),
+            'statusText' => 'Registro borrado con exito!',
+        );
+        echo json_encode($res);
+    }
+    if($tabla === 'procesadores' ){
+        $sql = "DELETE from $tabla where id_procesador = ? ";
+        $query = mysqli_prepare($mysqli , $sql);
+        $ok = mysqli_stmt_bind_param($query, 'i' , $id);
+        $ok = mysqli_stmt_execute($query);
+        mysqli_stmt_close($query);
+        $res = array (
+            'err' => false,
+            'status' => http_response_code(200),
+            'statusText' => 'Registro borrado con exito!',
+        );
+        echo json_encode($res);
+    }
+    if($tabla === 'tipo_sistema' ){
+        $sql = "DELETE from $tabla where id_tipo_sistema = ? ";
+        $query = mysqli_prepare($mysqli , $sql);
+        $ok = mysqli_stmt_bind_param($query, 'i' , $id);
+        $ok = mysqli_stmt_execute($query);
+        mysqli_stmt_close($query);
+        $res = array (
+            'err' => false,
+            'status' => http_response_code(200),
+            'statusText' => 'Registro borrado con exito!',
+        );
         echo json_encode($res);
     }
     if($tabla === 'dispositivo_electronico' ){
